исправлена работа с сессией

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -2,13 +2,14 @@ package jm.task.core.jdbc.dao;
 
 import jm.task.core.jdbc.model.User;
 import jm.task.core.jdbc.util.Util;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
-    Session session = Util.getSession();
 
     public UserDaoHibernateImpl() {
     }
@@ -22,30 +23,50 @@ public class UserDaoHibernateImpl implements UserDao {
     @Override
     public void dropUsersTable() {
         createDropCleanTable(UserDao.SQL_DROP, UserDao.PRINT_DROP);
-
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-        session.beginTransaction();
-        User user = new User(name, lastName, age);
-        session.save(user);
-        session.getTransaction().commit();
+        Transaction transaction = null;
+        try (Session session = Util.getSession()) {
+            transaction = session.beginTransaction();
+            User user = new User(name, lastName, age);
+            session.save(user);
+            transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
+            e.printStackTrace();
+        }
     }
 
     @Override
     public void removeUserById(long id) {
-        session.beginTransaction();
-        User user = session.get(User.class, id);
-        session.delete(user);
-        session.getTransaction().commit();
+        Transaction transaction = null;
+        try (Session session = Util.getSession()) {
+            transaction = session.beginTransaction();
+            User user = session.get(User.class, id);
+            if (user != null) {
+                session.delete(user);
+            }
+            transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
+            e.printStackTrace();
+        }
     }
 
     @Override
     public List<User> getAllUsers() {
-        session.beginTransaction();
-        List<User> users = session.createQuery("FROM User").getResultList();
-        session.getTransaction().commit();
+        List<User> users = new ArrayList<>();
+        Transaction transaction = null;
+        try (Session session = Util.getSession()) {
+            transaction = session.beginTransaction();
+            users = session.createQuery("FROM User").getResultList();
+            transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
+            e.printStackTrace();
+        }
         return users;
     }
 
@@ -55,9 +76,15 @@ public class UserDaoHibernateImpl implements UserDao {
     }
 
     private void createDropCleanTable(String sql, String print) {
-        Transaction transaction = session.beginTransaction();
-        session.createSQLQuery(sql).executeUpdate();
-        transaction.commit();
-        System.out.println(print);
+        Transaction transaction = null;
+        try (Session session = Util.getSession()) {
+            transaction = session.beginTransaction();
+            session.createSQLQuery(sql).executeUpdate();
+            transaction.commit();
+            System.out.println(print);
+        } catch (HibernateException e) {
+            if (transaction != null) transaction.rollback();
+            e.printStackTrace();
+        }
     }
 }
